Most seperation of concern is done. Still need to look at SeatingArea
