+++ package.json
@@ -22,7 +22,7 @@
     "cookie": "^0.5.0",
     "i": "^0.3.7",
     "jsonwebtoken": "^9.0.2",
-    "next": "^13.5.5",
+    "next": "^13.5.6",
     "react": "latest",
     "react-dom": "latest",
     "react-hook-form": "^7.46.2",
